Remove r4300 files. Move functions to r4300_core.
