+++ src/device/device.c
@@ -87,7 +87,7 @@ void poweron_device(struct device* dev)
 void run_device(struct device* dev)
 {
     /* device execution is driven by the r4300 */
-    r4300_execute();
+    run_r4300();
 }
 
 void stop_device(struct device* dev)
