+++ src/device/r4300/cp0.c
@@ -26,7 +26,6 @@
 #include "exception.h"
 #include "main/main.h"
 #include "new_dynarec/new_dynarec.h" /* for NEW_DYNAREC_ARM */
-#include "r4300.h"
 #include "recomp.h"
 
 #ifdef COMPARE_CORE
