+++ src/device/r4300/fpu.h
@@ -25,8 +25,6 @@
 #include <math.h>
 #include <stdint.h>
 
-#include "r4300.h"
-
 #ifdef _MSC_VER
   #define M64P_FPU_INLINE static __inline
   #include <float.h>
