+++ src/device/r4300/interupt.c
@@ -36,7 +36,6 @@
 #include "device/r4300/exception.h"
 #include "device/r4300/mi_controller.h"
 #include "device/r4300/new_dynarec/new_dynarec.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/r4300_core.h"
 #include "device/r4300/recomp.h"
 #include "device/rdp/rdp_core.h"
