+++ src/device/r4300/mi_controller.c
@@ -25,7 +25,6 @@
 
 #include "cp0.h"
 #include "interupt.h"
-#include "r4300.h"
 #include "r4300_core.h"
 
 static int update_mi_init_mode(uint32_t* mi_init_mode, uint32_t w)
