+++ src/device/r4300/r4300_core.c
@@ -20,13 +20,24 @@
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
 #include "r4300_core.h"
-
 #include "cached_interp.h"
+#if defined(COUNT_INSTR)
+#include "instr_counters.h"
+#endif
 #include "mi_controller.h"
 #include "new_dynarec/new_dynarec.h" /* for NEW_DYNAREC_ARM */
-#include "r4300.h"
+#include "pure_interp.h"
 #include "recomp.h"
+
+#include "api/callbacks.h"
+#include "api/debugger.h"
+#include "api/m64p_types.h"
+#ifdef DBG
+#include "debugger/dbg_debugger.h"
+#include "debugger/dbg_types.h"
+#endif
 #include "main/main.h"
+
 
 #include <string.h>
 
@@ -97,6 +108,119 @@ void poweron_r4300(struct r4300_core* r4300)
 
     /* setup mi */
     poweron_mi(&r4300->mi);
+}
+
+
+#if !defined(NO_ASM)
+static void dynarec_setup_code(void)
+{
+   // The dynarec jumps here after we call dyna_start and it prepares
+   // Here we need to prepare the initial code block and jump to it
+   jump_to(UINT32_C(0xa4000040));
+
+   // Prevent segfault on failed jump_to
+   if (!g_dev.r4300.cached_interp.actual->block || !g_dev.r4300.cached_interp.actual->code)
+      dyna_stop();
+}
+#endif
+
+void run_r4300(void)
+{
+#if (defined(DYNAREC) && defined(PROFILE_R4300))
+    unsigned int i;
+#endif
+
+    g_dev.r4300.current_instruction_table = cached_interpreter_table;
+
+    *r4300_stop() = 0;
+    g_rom_pause = 0;
+
+    /* clear instruction counters */
+#if defined(COUNT_INSTR)
+    memset(instr_count, 0, 131*sizeof(instr_count[0]));
+#endif
+
+    /* XXX: might go to r4300_poweron / soft_reset ? */
+    g_dev.r4300.cp0.last_addr = 0xa4000040;
+    *r4300_cp0_next_interrupt() = 624999;
+    init_interupt();
+
+    if (g_dev.r4300.emumode == EMUMODE_PURE_INTERPRETER)
+    {
+        DebugMessage(M64MSG_INFO, "Starting R4300 emulator: Pure Interpreter");
+        g_dev.r4300.emumode = EMUMODE_PURE_INTERPRETER;
+        pure_interpreter();
+    }
+#if defined(DYNAREC)
+    else if (g_dev.r4300.emumode >= 2)
+    {
+        DebugMessage(M64MSG_INFO, "Starting R4300 emulator: Dynamic Recompiler");
+        g_dev.r4300.emumode = EMUMODE_DYNAREC;
+        init_blocks();
+
+#ifdef NEW_DYNAREC
+        new_dynarec_init();
+        new_dyna_start();
+        new_dynarec_cleanup();
+#else
+        dyna_start(dynarec_setup_code);
+        (*r4300_pc_struct())++;
+#endif
+#if defined(PROFILE_R4300)
+        g_dev.r4300.recomp.pfProfile = fopen("instructionaddrs.dat", "ab");
+        for (i=0; i<0x100000; i++)
+            if (g_dev.r4300.cached_interp.invalid_code[i] == 0 && g_dev.r4300.cached_interp.blocks[i] != NULL && g_dev.r4300.cached_interp.blocks[i]->code != NULL && g_dev.r4300.cached_interp.blocks[i]->block != NULL)
+            {
+                unsigned char *x86addr;
+                int mipsop;
+                // store final code length for this block
+                mipsop = -1; /* -1 == end of x86 code block */
+                x86addr = g_dev.r4300.cached_interp.blocks[i]->code + g_dev.r4300.cached_interp.blocks[i]->code_length;
+                if (fwrite(&mipsop, 1, 4, g_dev.r4300.recomp.pfProfile) != 4 ||
+                    fwrite(&x86addr, 1, sizeof(char *), g_dev.r4300.recomp.pfProfile) != sizeof(char *))
+                    DebugMessage(M64MSG_ERROR, "Error writing R4300 instruction address profiling data");
+            }
+        fclose(g_dev.r4300.recomp.pfProfile);
+        g_dev.r4300.recomp.pfProfile = NULL;
+#endif
+        free_blocks();
+    }
+#endif
+    else /* if (g_dev.r4300.emumode == EMUMODE_INTERPRETER) */
+    {
+        DebugMessage(M64MSG_INFO, "Starting R4300 emulator: Cached Interpreter");
+        g_dev.r4300.emumode = EMUMODE_INTERPRETER;
+        init_blocks();
+        jump_to(UINT32_C(0xa4000040));
+
+        /* Prevent segfault on failed jump_to */
+        if (!g_dev.r4300.cached_interp.actual->block)
+            return;
+
+        g_dev.r4300.cp0.last_addr = *r4300_pc();
+        while (!*r4300_stop())
+        {
+#ifdef COMPARE_CORE
+            if ((*r4300_pc_struct())->ops == cached_interpreter_table.FIN_BLOCK && ((*r4300_pc_struct())->addr < 0x80000000 || (*r4300_pc_struct())->addr >= 0xc0000000))
+                virtual_to_physical_address((*r4300_pc_struct())->addr, 2);
+            CoreCompareCallback();
+#endif
+#ifdef DBG
+            if (g_DebuggerActive) update_debugger((*r4300_pc_struct())->addr);
+#endif
+            (*r4300_pc_struct())->ops();
+        }
+
+        free_blocks();
+    }
+
+    DebugMessage(M64MSG_INFO, "R4300 emulator finished.");
+
+    /* print instruction counts */
+#if defined(COUNT_INSTR)
+    if (g_dev.r4300.emumode == EMUMODE_DYNAREC)
+        instr_counters_print();
+#endif
 }
 
 int64_t* r4300_regs(void)
@@ -188,6 +312,24 @@ void invalidate_r4300_cached_code(uint32_t address, size_t size)
         }
     }
 }
+
+
+void generic_jump_to(uint32_t address)
+{
+   if (g_dev.r4300.emumode == EMUMODE_PURE_INTERPRETER)
+      *r4300_pc() = address;
+   else {
+#ifdef NEW_DYNAREC
+      if (g_dev.r4300.emumode == EMUMODE_DYNAREC)
+         g_dev.r4300.cp0.last_addr = pcaddr;
+      else
+         jump_to(address);
+#else
+      jump_to(address);
+#endif
+   }
+}
+
 
 /* XXX: not really a good interface but it gets the job done... */
 void savestates_load_set_pc(uint32_t pc)
