+++ src/device/r4300/r4300_core.h
@@ -170,6 +170,8 @@ struct r4300_core
 void init_r4300(struct r4300_core* r4300, unsigned int emumode, unsigned int count_per_op, int no_compiled_jump);
 void poweron_r4300(struct r4300_core* r4300);
 
+void run_r4300(void);
+
 int64_t* r4300_regs(void);
 int64_t* r4300_mult_hi(void);
 int64_t* r4300_mult_lo(void);
