+++ src/device/r4300/x86/assemble.c
@@ -25,7 +25,6 @@
 #include "api/callbacks.h"
 #include "api/m64p_types.h"
 #include "assemble.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
 #include "osal/preproc.h"
