+++ src/device/r4300/x86/gcop1_d.c
@@ -25,7 +25,6 @@
 #include "interpret.h"
 #include "device/r4300/cp1.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomph.h"
 #include "main/main.h"
