+++ src/device/r4300/x86/gcop1_s.c
@@ -26,7 +26,6 @@
 #include "device/r4300/cp1.h"
 #include "device/r4300/macros.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomph.h"
 #include "main/main.h"
