+++ src/device/r4300/x86/gregimm.c
@@ -27,7 +27,6 @@
 #include "device/r4300/cached_interp.h"
 #include "device/r4300/macros.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
 #include "main/main.h"
