+++ src/device/r4300/x86/gtlb.c
@@ -24,7 +24,6 @@
 #include "assemble.h"
 #include "device/r4300/cached_interp.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomph.h"
 
 void gentlbwi(void)
