+++ src/device/r4300/x86/regcache.c
@@ -22,7 +22,6 @@
 #include <stdio.h>
 
 #include "regcache.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
 #include "main/main.h"
