+++ src/device/r4300/x86_64/gcop0.c
@@ -25,7 +25,6 @@
 #include "assemble.h"
 #include "device/r4300/cached_interp.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
