+++ src/device/r4300/x86_64/gcop1_w.c
@@ -26,7 +26,6 @@
 #include "interpret.h"
 #include "device/r4300/cp1.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
