+++ src/device/r4300/x86_64/gr4300.c
@@ -33,7 +33,6 @@
 #include "device/r4300/exception.h"
 #include "device/r4300/interupt.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
 #include "main/main.h"
