+++ src/device/r4300/x86_64/gtlb.c
@@ -25,7 +25,6 @@
 #include "assemble.h"
 #include "device/r4300/cached_interp.h"
 #include "device/r4300/ops.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomph.h"
 
 #if defined(COUNT_INSTR)
