+++ src/device/r4300/x86_64/regcache.c
@@ -25,7 +25,6 @@
 
 #include "api/callbacks.h"
 #include "api/m64p_types.h"
-#include "device/r4300/r4300.h"
 #include "device/r4300/recomp.h"
 #include "device/r4300/recomph.h"
 #include "device/r4300/x86_64/assemble.h"
